requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==2.1.2
django-debug-toolbar==1.10.1
django-registration==3.0
idna==2.7
kombu==4.2.1
psycopg2==2.7.5
pytz==2018.5
redis==2.10.6
requests==2.19.1
sqlparse==0.2.4
urllib3==1.23
vine==1.1.4
